fix: change double to int timestamp query

+++ Communication/Packets/Incoming/Moderation/GetModeratorUserChatlogEvent.cs
@@ -30,7 +30,7 @@ internal sealed class GetModeratorUserChatlogEvent : IPacketEvent
             {
                 foreach (DataRow row in table.Rows)
                 {
-                    sortedMessages.Add(new ChatlogEntry(Convert.ToInt32(row["user_id"]), row["user_name"].ToString(), Convert.ToInt32(row["room_id"]), row["type"].ToString() + row["message"].ToString(), (double)row["timestamp"]));
+                    sortedMessages.Add(new ChatlogEntry(Convert.ToInt32(row["user_id"]), row["user_name"].ToString(), Convert.ToInt32(row["room_id"]), row["type"].ToString() + row["message"].ToString(), (int)row["timestamp"]));
                 }
             }
